auth-1

+++ src/controllers/userController.js
@@ -1,13 +1,7 @@
 const jwt = require("jsonwebtoken");
 const userModel = require("../models/userModel");
 
-/*
-  Read all the comments multiple times to understand why we are doing what we are doing in login api and getUserData api
-*/
 const createUser = async function (abcd, xyz) {
-  //You can name the req, res objects anything.
-  //but the first parameter is always the request 
-  //the second parameter is always the response
   let data = abcd.body;
   let savedData = await userModel.create(data);
   console.log(abcd.newAtribute);
@@ -25,12 +19,6 @@ const loginUser = async function (req, res) {
       msg: "username or the password is not corerct",
     });
 
-  // Once the login is successful, create the jwt token with sign function
-  // Sign function has 2 inputs:
-  // Input 1 is the payload or the object containing data to be set in token
-  // The decision about what data to put in token depends on the business requirement
-  // Input 2 is the secret (This is basically a fixed value only set at the server. This value should be hard to guess)
-  // The same secret will be used to decode tokens 
   let token = jwt.sign(
     {
       userId: user._id.toString(),
@@ -52,15 +40,6 @@ const getUserData = async function (req, res) {
 
   console.log(token);
 
-  // If a token is present then decode the token with verify function
-  // verify takes two inputs:
-  // Input 1 is the token to be decoded
-  // Input 2 is the same secret with which the token was generated
-  // Check the value of the decoded token yourself
-
-  // Decoding requires the secret again. 
-  // A token can only be decoded successfully if the same secret was used to create(sign) that token.
-  // And because this token is only known to the server, it can be assumed that if a token is decoded at server then this token must have been issued by the same server in past.
   let decodedToken = jwt.verify(token, "functionup-plutonium");
   if (!decodedToken)
     return res.send({ status: false, msg: "token is invalid" });
@@ -69,20 +48,23 @@ const getUserData = async function (req, res) {
   let userDetails = await userModel.findById(userId);
   if (!userDetails)
     return res.send({ status: false, msg: "No such user exists" });
-
+let limit = req.params.userId
+let acces=decodedToken.userId
+if(limit != acces) return res.send({status:false, msg:"you can't see other user personal data"})
   res.send({ status: true, data: userDetails });
   // Note: Try to see what happens if we change the secret while decoding the token
 };
 
 const updateUser = async function (req, res) {
-  // Do the same steps here:
-  // Check if the token is present
-  // Check if the token present is a valid token
-  // Return a different error message in both these cases
-
   let userId = req.params.userId;
+  let token = req.headers["x-auth-token"]
+  if (!token) return res.send({ status: false, msg: "token must bhi present" })
+  let decode = jwt.verify(token, 'functionup-plutonium')
+  if (!decode) return res.send({ status: false, msg: "token input is not correct" })
+  let limit = req.params.userId
+  let acces = decode.userId
+  if (limit != acces) return res.send({ status: false, msg: "you cant update other person data" })
   let user = await userModel.findById(userId);
-  //Return an error if no user with the given id exists in the db
   if (!user) {
     return res.send("No such user exists");
   }
@@ -92,19 +74,26 @@ const updateUser = async function (req, res) {
   res.send({ status: updatedUser, data: updatedUser });
 };
 
-const deleteUser = async function (req ,res){
+const deleteUser = async function (req, res) {
   let userId = req.params.userId;
-let user = await userModel.findById(userId)
-if(!user){
-  res.send("no such user exists")
-}
-let deleteData = req.body;
-let del = await userModel.findOneAndDelete({_id:userId},deleteData)
-res.send({status : del , data : del})
+  let token = req.headers["x-auth-token"]
+  if(!token) return res.send({status:false , msg:"token must bhi present"})
+  let decode =jwt.verify(token , 'function-plutonium')
+  if(!decode) return res.send({status:false , msg:"token is invalid"})
+  let userlimit = req.params.userId
+  let acces =decode.userId
+  if(limit != acces)return res.send({status : false ,msg: "you can't delete other user data "})
+  let user = await userModel.findById(userId)
+  if (!user) {
+    res.send("no such user exists")
+  }
+  let deleteData = req.body;
+  let del = await userModel.findOneAndDelete({ _id: userId }, deleteData)
+  res.send({ status: del, data: del })
 }
 
 module.exports.createUser = createUser;
 module.exports.getUserData = getUserData;
 module.exports.updateUser = updateUser;
 module.exports.loginUser = loginUser;
-module.exports.deleteUser=deleteUser;
+module.exports.deleteUser = deleteUser;
